fonts: Use variable font for Roboto Regular
I'm not sure why Google decided to revert to a static variant of *only*
the regular non-italic and non-condensed weight, but using the default
"Roboto" font for the lock screen clock (i.e. AOSP default) is broken
with the change:
    - Clock is too bold when notifications are present (because SystemUI
      loads the regular font but can't change the weight to be less
      bold)
    - Weight animation between lock screen and AOD is broken (because
      weight can't be animated without the variable font)
    - AOD clock font is too bold (could cause burn-in)

Switch back to the variable font in order to fix the clock issues.

Change-Id: Ia2e5ef52e4edb5d4275f6d5faf8c74266b784762
Reviewed-on: https://review.statixos.com/c/android_frameworks_base/+/5786
Reviewed-by: Anay Wadhera <anay1018@gmail.com>
Reviewed-by: Sourajit Karmakar <sourajit@live.com>
Tested-by: Anay Wadhera <anay1018@gmail.com>
Tested-by: Sourajit Karmakar <sourajit@live.com>
Signed-off-by: Joey Huab <joey@evolution-x.org>